moving getter functions into a new file
